Switched to a 16-bit _sessionID, not tested yet.

+++ ATEM.h
@@ -48,7 +48,7 @@ class ATEM
 	IPAddress _switcherIP;		// IP address of the switcher
 	boolean _serialOutput;		// If set, the library will print status/debug information to the Serial object
 
-	uint8_t _sessionID;					// Used internally for storing packet size during communication
+	uint16_t _sessionID;					// Used internally for storing packet size during communication
 	uint16_t _lastRemotePacketID;		// The most recent Remote Packet Id from switcher
 	uint8_t _packetBuffer[96];   			// Buffer for storing segments of the packets from ATEM and creating answer packets.
 	uint16_t _cmdLength;					// Used when parsing packets
